+++ examples/Kemper Profiler Player/MIDICaptain 10/Arek77/inputs.py
@@ -1,84 +1,110 @@
 ##############################################################################################################################################
 # 
-# Definition of actions for switches
+# Definition of actions for switches DB Version
 #
 ##############################################################################################################################################
  
 from pyswitch.hardware.Hardware import Hardware
 
+# from pyswitch.misc import Colors
+
 from pyswitch.clients.kemper import KemperEffectSlot
+from display import DISPLAY_FOOTER_1, DISPLAY_FOOTER_2, DISPLAY_HEADER_1, DISPLAY_HEADER_2
+
 from pyswitch.clients.kemper.actions.rig_select import RIG_SELECT, RIG_SELECT_DISPLAY_TARGET_RIG
 from pyswitch.clients.kemper.actions.bank_select import BANK_SELECT
 from pyswitch.clients.kemper.actions.effect_state import EFFECT_STATE
+from pyswitch.clients.kemper.actions.morph import MORPH_DISPLAY
+from pyswitch.clients.kemper.actions.tuner import TUNER_MODE
 
-from display import DISPLAY_HEADER_1, DISPLAY_HEADER_2, DISPLAY_FOOTER_1, DISPLAY_FOOTER_2
-
-##############################################################################################################################################
 
 # Defines the switch assignments
-Switches = [
+Inputs = [
 
     # Switch 1
     {
         "assignment": Hardware.PA_MIDICAPTAIN_10_SWITCH_1,
         "actions": [
+            EFFECT_STATE(
+                slot_id = KemperEffectSlot.EFFECT_SLOT_ID_A,
+                display = DISPLAY_HEADER_1
+            )
+        ],
+        "actionsHold": [
             BANK_SELECT(
                 bank = 1,
-                preselect = True,
                 display_mode = RIG_SELECT_DISPLAY_TARGET_RIG
-            )         
-        ]
+            )
+        ]    
     },
 
     # Switch 2
     {
         "assignment": Hardware.PA_MIDICAPTAIN_10_SWITCH_2,
         "actions": [
+            EFFECT_STATE(
+                slot_id = KemperEffectSlot.EFFECT_SLOT_ID_B,
+                display = DISPLAY_HEADER_2          
+            )
+        ],
+        "actionsHold": [
             BANK_SELECT(
                 bank = 2,
-                preselect = True,
+                bank_off = 4,
                 display_mode = RIG_SELECT_DISPLAY_TARGET_RIG
-            )     
-        ]
+            )
+        ]    
     },
 
     # Switch 3
     {
         "assignment": Hardware.PA_MIDICAPTAIN_10_SWITCH_3,
         "actions": [
+            EFFECT_STATE(
+                slot_id = KemperEffectSlot.EFFECT_SLOT_ID_DLY,
+                display = DISPLAY_FOOTER_1
+            )
+        ],
+        "actionsHold": [
             BANK_SELECT(
                 bank = 3,
-                preselect = True,
                 display_mode = RIG_SELECT_DISPLAY_TARGET_RIG
-            )     
-        ]
+            )
+        ]    
     },
-
+    
     # Switch 4
     {
         "assignment": Hardware.PA_MIDICAPTAIN_10_SWITCH_4,
         "actions": [
+                EFFECT_STATE(
+                slot_id = KemperEffectSlot.EFFECT_SLOT_ID_REV,
+                display = DISPLAY_FOOTER_2
+            )
+        ],
+        "actionsHold": [
             BANK_SELECT(
                 bank = 4,
-                preselect = True,
                 display_mode = RIG_SELECT_DISPLAY_TARGET_RIG
-            )     
-        ]
+            )
+        ]    
     },
 
     # Switch up
     {
         "assignment": Hardware.PA_MIDICAPTAIN_10_SWITCH_UP,
         "actions": [
+            TUNER_MODE()
+        ],
+        "actionsHold": [
             BANK_SELECT(
                 bank = 5,
-                preselect = True,
                 display_mode = RIG_SELECT_DISPLAY_TARGET_RIG
-            )     
-        ]
+            )
+        ]    
     },
 
-#################################################################
+    # ########################################################################################
 
     # Switch A
     {
@@ -86,32 +112,36 @@ Switches = [
         "actions": [
             RIG_SELECT(
                 rig = 1,
-                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG
-            )   
+                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG,
+                #color = Colors.YELLOW
+            ),
+            MORPH_DISPLAY()
         ],
         "actionsHold": [
-            EFFECT_STATE(
-                slot_id = KemperEffectSlot.EFFECT_SLOT_ID_A,
-                display = DISPLAY_HEADER_1
-            )   
-        ]
+            BANK_SELECT(
+                bank = 6,
+                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG
+            )
+        ]    
     },
-    
+
     # Switch B
     {
         "assignment": Hardware.PA_MIDICAPTAIN_10_SWITCH_B,
         "actions": [
             RIG_SELECT(
                 rig = 2,
-                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG
-            )   
+                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG,
+                #color = Colors.BLUE
+            ),
+            MORPH_DISPLAY()
         ],
         "actionsHold": [
-            EFFECT_STATE(
-                slot_id = KemperEffectSlot.EFFECT_SLOT_ID_B,
-                display = DISPLAY_HEADER_2
-            )   
-        ]
+            BANK_SELECT(
+                bank = 7,
+                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG
+            )
+        ]    
     },
 
     # Switch C
@@ -120,15 +150,17 @@ Switches = [
         "actions": [
             RIG_SELECT(
                 rig = 3,
-                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG
-            )   
+                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG,
+                #color = Colors.ORANGE
+            ),
+            MORPH_DISPLAY()
         ],
         "actionsHold": [
-            EFFECT_STATE(
-                slot_id = KemperEffectSlot.EFFECT_SLOT_ID_C,
-                display = DISPLAY_FOOTER_1
-            )   
-        ]
+            BANK_SELECT(
+                bank = 8,
+                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG
+            )
+        ]    
     },
 
     # Switch D
@@ -137,15 +169,17 @@ Switches = [
         "actions": [
             RIG_SELECT(
                 rig = 4,
-                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG
-            )   
+                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG,
+                #color = Colors.RED
+            ),
+            MORPH_DISPLAY()
         ],
         "actionsHold": [
-            EFFECT_STATE(
-                slot_id = KemperEffectSlot.EFFECT_SLOT_ID_D,
-                display = DISPLAY_FOOTER_2
-            )   
-        ]
+            BANK_SELECT(
+                bank = 9,
+                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG
+            )
+        ]    
     },
 
     # Switch down
@@ -154,13 +188,16 @@ Switches = [
         "actions": [
             RIG_SELECT(
                 rig = 5,
+                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG,
+                #color = Colors.GREEN
+            ),
+            MORPH_DISPLAY()
+        ],
+        "actionsHold": [
+            BANK_SELECT(
+                bank = 10,
                 display_mode = RIG_SELECT_DISPLAY_TARGET_RIG
-            )   
-        ]
+            )
+        ]    
     },
 ]
-
-#####################################################################################
-
-# Expression pedals and other inputs
-Inputs = None