+++ examples/Kemper Profiler Player/MIDICaptain 10/Bank Preselect and FX States/inputs.py
@@ -1,0 +1,161 @@
+##############################################################################################################################################
+# 
+# Definition of actions for switches
+#
+##############################################################################################################################################
+ 
+from pyswitch.hardware.Hardware import Hardware
+
+from pyswitch.clients.kemper import KemperEffectSlot
+from pyswitch.clients.kemper.actions.rig_select import RIG_SELECT, RIG_SELECT_DISPLAY_TARGET_RIG
+from pyswitch.clients.kemper.actions.bank_select import BANK_SELECT
+from pyswitch.clients.kemper.actions.effect_state import EFFECT_STATE
+
+from display import DISPLAY_HEADER_1, DISPLAY_HEADER_2, DISPLAY_FOOTER_1, DISPLAY_FOOTER_2
+
+##############################################################################################################################################
+
+# Defines the switch assignments
+Inputs = [
+
+    # Switch 1
+    {
+        "assignment": Hardware.PA_MIDICAPTAIN_10_SWITCH_1,
+        "actions": [
+            BANK_SELECT(
+                bank = 1,
+                preselect = True,
+                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG
+            )         
+        ]
+    },
+
+    # Switch 2
+    {
+        "assignment": Hardware.PA_MIDICAPTAIN_10_SWITCH_2,
+        "actions": [
+            BANK_SELECT(
+                bank = 2,
+                preselect = True,
+                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG
+            )     
+        ]
+    },
+
+    # Switch 3
+    {
+        "assignment": Hardware.PA_MIDICAPTAIN_10_SWITCH_3,
+        "actions": [
+            BANK_SELECT(
+                bank = 3,
+                preselect = True,
+                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG
+            )     
+        ]
+    },
+
+    # Switch 4
+    {
+        "assignment": Hardware.PA_MIDICAPTAIN_10_SWITCH_4,
+        "actions": [
+            BANK_SELECT(
+                bank = 4,
+                preselect = True,
+                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG
+            )     
+        ]
+    },
+
+    # Switch up
+    {
+        "assignment": Hardware.PA_MIDICAPTAIN_10_SWITCH_UP,
+        "actions": [
+            BANK_SELECT(
+                bank = 5,
+                preselect = True,
+                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG
+            )     
+        ]
+    },
+
+#################################################################
+
+    # Switch A
+    {
+        "assignment": Hardware.PA_MIDICAPTAIN_10_SWITCH_A,
+        "actions": [
+            RIG_SELECT(
+                rig = 1,
+                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG
+            )   
+        ],
+        "actionsHold": [
+            EFFECT_STATE(
+                slot_id = KemperEffectSlot.EFFECT_SLOT_ID_A,
+                display = DISPLAY_HEADER_1
+            )   
+        ]
+    },
+    
+    # Switch B
+    {
+        "assignment": Hardware.PA_MIDICAPTAIN_10_SWITCH_B,
+        "actions": [
+            RIG_SELECT(
+                rig = 2,
+                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG
+            )   
+        ],
+        "actionsHold": [
+            EFFECT_STATE(
+                slot_id = KemperEffectSlot.EFFECT_SLOT_ID_B,
+                display = DISPLAY_HEADER_2
+            )   
+        ]
+    },
+
+    # Switch C
+    {
+        "assignment": Hardware.PA_MIDICAPTAIN_10_SWITCH_C,
+        "actions": [
+            RIG_SELECT(
+                rig = 3,
+                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG
+            )   
+        ],
+        "actionsHold": [
+            EFFECT_STATE(
+                slot_id = KemperEffectSlot.EFFECT_SLOT_ID_C,
+                display = DISPLAY_FOOTER_1
+            )   
+        ]
+    },
+
+    # Switch D
+    {
+        "assignment": Hardware.PA_MIDICAPTAIN_10_SWITCH_D,
+        "actions": [
+            RIG_SELECT(
+                rig = 4,
+                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG
+            )   
+        ],
+        "actionsHold": [
+            EFFECT_STATE(
+                slot_id = KemperEffectSlot.EFFECT_SLOT_ID_D,
+                display = DISPLAY_FOOTER_2
+            )   
+        ]
+    },
+
+    # Switch down
+    {
+        "assignment": Hardware.PA_MIDICAPTAIN_10_SWITCH_DOWN,
+        "actions": [
+            RIG_SELECT(
+                rig = 5,
+                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG
+            )   
+        ]
+    },
+]