+++ examples/Kemper Profiler Player/MIDICaptain Mini 6/FX States and Rig Volume plus Tuner 2/inputs.py
@@ -16,7 +16,7 @@ from pyswitch.clients.kemper.actions.effect_state import EFFECT_STATE
 
 
 # Defines the switch assignments
-Switches = [
+Inputs = [
 
     # Switch 1
     {
@@ -84,8 +84,3 @@ Switches = [
         ]
     }
 ]
-
-######################################################################
-
-# Expression pedals and other inputs
-Inputs = None