+++ examples/Kemper Profiler Player/MIDICaptain Nano 4/Bank Up Down/inputs.py
@@ -6,25 +6,21 @@
  
 from pyswitch.hardware.Hardware import Hardware
 
-from pyswitch.misc import Colors
-
 from pyswitch.clients.kemper import KemperEffectSlot
-from display import DISPLAY_FOOTER_1, DISPLAY_FOOTER_2, DISPLAY_HEADER_1, DISPLAY_HEADER_2
+from display import DISPLAY_HEADER_1, DISPLAY_HEADER_2, DISPLAY_FOOTER_1, DISPLAY_FOOTER_2
 
-from pyswitch.clients.kemper.actions.tempo import TAP_TEMPO, SHOW_TEMPO
-from pyswitch.clients.kemper.actions.tuner import TUNER_MODE
-from pyswitch.clients.kemper.actions.morph import MORPH_BUTTON
+from pyswitch.clients.kemper.actions.bank_up_down import BANK_UP, BANK_DOWN
 from pyswitch.clients.kemper.actions.effect_state import EFFECT_STATE
 
 
 # Defines the switch assignments
-Switches = [
+Inputs = [
 
     # Switch 1
     {
         "assignment": Hardware.PA_MIDICAPTAIN_NANO_SWITCH_1,
         "actions": [
-            TUNER_MODE(
+            BANK_DOWN(
                 display = DISPLAY_HEADER_1
             )                         
         ]
@@ -34,12 +30,9 @@ Switches = [
     {
         "assignment": Hardware.PA_MIDICAPTAIN_NANO_SWITCH_2,
         "actions": [
-            TAP_TEMPO(
-                display = DISPLAY_HEADER_2,
-                color = Colors.LIGHT_GREEN,
-                use_leds = False
-            ),
-            SHOW_TEMPO()    # Shows beats with the LED(s)   
+            BANK_UP(
+                display = DISPLAY_HEADER_2
+            )
         ]
     },
 
@@ -48,7 +41,7 @@ Switches = [
         "assignment": Hardware.PA_MIDICAPTAIN_NANO_SWITCH_A,
         "actions": [
             EFFECT_STATE(
-                slot_id = KemperEffectSlot.EFFECT_SLOT_ID_A,
+                slot_id = KemperEffectSlot.EFFECT_SLOT_ID_DLY,
                 display = DISPLAY_FOOTER_1
             )
         ]
@@ -58,15 +51,10 @@ Switches = [
     {
         "assignment": Hardware.PA_MIDICAPTAIN_NANO_SWITCH_B,
         "actions": [
-            MORPH_BUTTON(
-                display = DISPLAY_FOOTER_2,
-                #color = "kemper"
+            EFFECT_STATE(
+                slot_id = KemperEffectSlot.EFFECT_SLOT_ID_REV,
+                display = DISPLAY_FOOTER_2
             )
         ]
     }
 ]
-
-######################################################################
-
-# Expression pedals and other inputs
-Inputs = None