+++ examples/Kemper Profiler Player/MIDICaptain Nano 4/FX States and Bank Change/inputs.py
@@ -16,7 +16,7 @@ from pyswitch.clients.kemper.actions.effect_state import EFFECT_STATE
 
 
 # Defines the switch assignments
-Switches = [
+Inputs = [
 
     # Switch 1
     {
@@ -68,8 +68,3 @@ Switches = [
         ]
     },
 ]
-
-######################################################################
-
-# Expression pedals and other inputs
-Inputs = None