+++ examples/Kemper Profiler Player/MIDICaptain Nano 4/Mod Speed and Amp State/inputs.py
@@ -9,52 +9,56 @@ from pyswitch.hardware.Hardware import Hardware
 from pyswitch.misc import Colors
 
 from pyswitch.clients.kemper import KemperEffectSlot
-
 from display import DISPLAY_HEADER_1, DISPLAY_HEADER_2, DISPLAY_FOOTER_1, DISPLAY_FOOTER_2
 
-from pyswitch.clients.kemper.actions.morph import MORPH_DISPLAY
+from pyswitch.clients.kemper.actions.binary_switch import BINARY_SWITCH
 from pyswitch.clients.kemper.actions.tuner import TUNER_MODE
 from pyswitch.clients.kemper.actions.effect_state import EFFECT_STATE
-from pyswitch.clients.kemper.actions.rig_select import RIG_SELECT, RIG_SELECT_DISPLAY_TARGET_RIG
+
+from pyswitch.clients.kemper.mappings.rotary import MAPPING_ROTARY_SPEED
+from pyswitch.clients.kemper.mappings.amp import MAPPING_AMP_STATE
 
 
 # Defines the switch assignments
-Switches = [
+Inputs = [
 
     # Switch 1
     {
         "assignment": Hardware.PA_MIDICAPTAIN_NANO_SWITCH_1,
         "actions": [
-            TUNER_MODE(
+            EFFECT_STATE(
+                slot_id = KemperEffectSlot.EFFECT_SLOT_ID_A,
                 display = DISPLAY_HEADER_1
-            ),
-            MORPH_DISPLAY()        
+            )
+        ],
+        "actionsHold": [
+            TUNER_MODE()
         ]
     },
 
     # Switch 2
     {
         "assignment": Hardware.PA_MIDICAPTAIN_NANO_SWITCH_2,
-        "actions": [   
+        "actions": [
             EFFECT_STATE(
-                slot_id = KemperEffectSlot.EFFECT_SLOT_ID_X,
+                slot_id = KemperEffectSlot.EFFECT_SLOT_ID_B,
                 display = DISPLAY_HEADER_2
             )
-        ]               
+        ]
     },
 
     # Switch A
     {
         "assignment": Hardware.PA_MIDICAPTAIN_NANO_SWITCH_A,
         "actions": [
-            RIG_SELECT(
-                rig = 3,
-                rig_off = 1,
+            BINARY_SWITCH(
+                mapping = MAPPING_ROTARY_SPEED(
+                    slot_id = KemperEffectSlot.EFFECT_SLOT_ID_A
+                ),
                 display = DISPLAY_FOOTER_1,
-                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG,
-                color = Colors.PINK,
-                text = "Synth-3"
-            )            
+                color = Colors.DARK_BLUE,
+                text = "A: Fast"
+            )
         ]
     },
     
@@ -62,18 +66,12 @@ Switches = [
     {
         "assignment": Hardware.PA_MIDICAPTAIN_NANO_SWITCH_B,
         "actions": [
-            RIG_SELECT(
-                rig = 2,
-                rig_off = 1,
+            BINARY_SWITCH(
+                mapping = MAPPING_AMP_STATE(),
                 display = DISPLAY_FOOTER_2,
-                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG,
-                text = "Lead-2"
+                color = Colors.GRAY,
+                text = "Amp"
             )
         ]
     }
 ]
-
-######################################################################
-
-# Expression pedals and other inputs
-Inputs = None