+++ examples/Kemper Profiler Player/MIDICaptain Nano 4/Tuner and Tap Tempo and Morph Button/inputs.py
@@ -7,26 +7,24 @@
 from pyswitch.hardware.Hardware import Hardware
 
 from pyswitch.misc import Colors
-from pyswitch.controller.actions import PushButtonAction
-from pyswitch.controller.callbacks import BinaryParameterCallback
 
 from pyswitch.clients.kemper import KemperEffectSlot
-from display import DISPLAY_HEADER_1, DISPLAY_HEADER_2, DISPLAY_FOOTER_1, DISPLAY_FOOTER_2
+from display import DISPLAY_FOOTER_1, DISPLAY_FOOTER_2, DISPLAY_HEADER_1, DISPLAY_HEADER_2
 
 from pyswitch.clients.kemper.actions.tempo import TAP_TEMPO, SHOW_TEMPO
+from pyswitch.clients.kemper.actions.tuner import TUNER_MODE
+from pyswitch.clients.kemper.actions.morph import MORPH_BUTTON
 from pyswitch.clients.kemper.actions.effect_state import EFFECT_STATE
-from pyswitch.clients.kemper.mappings.freeze import MAPPING_FREEZE
 
 
 # Defines the switch assignments
-Switches = [
+Inputs = [
 
     # Switch 1
     {
         "assignment": Hardware.PA_MIDICAPTAIN_NANO_SWITCH_1,
         "actions": [
-            EFFECT_STATE(
-                slot_id = KemperEffectSlot.EFFECT_SLOT_ID_A,
+            TUNER_MODE(
                 display = DISPLAY_HEADER_1
             )                         
         ]
@@ -36,10 +34,12 @@ Switches = [
     {
         "assignment": Hardware.PA_MIDICAPTAIN_NANO_SWITCH_2,
         "actions": [
-            EFFECT_STATE(
-                slot_id = KemperEffectSlot.EFFECT_SLOT_ID_B,
-                display = DISPLAY_HEADER_2
-            )
+            TAP_TEMPO(
+                display = DISPLAY_HEADER_2,
+                color = Colors.LIGHT_GREEN,
+                use_leds = False
+            ),
+            SHOW_TEMPO()    # Shows beats with the LED(s)   
         ]
     },
 
@@ -47,16 +47,10 @@ Switches = [
     {
         "assignment": Hardware.PA_MIDICAPTAIN_NANO_SWITCH_A,
         "actions": [
-            PushButtonAction({
-                "callback": BinaryParameterCallback(
-                    mapping = MAPPING_FREEZE(KemperEffectSlot.EFFECT_SLOT_ID_REV),
-                    text = "RvFreeze",
-                    color = Colors.DARK_GREEN
-                ),
-                "mode": PushButtonAction.MOMENTARY,
-                "display": DISPLAY_FOOTER_1,
-                "useSwitchLeds": True
-            })
+            EFFECT_STATE(
+                slot_id = KemperEffectSlot.EFFECT_SLOT_ID_A,
+                display = DISPLAY_FOOTER_1
+            )
         ]
     },
     
@@ -64,17 +58,10 @@ Switches = [
     {
         "assignment": Hardware.PA_MIDICAPTAIN_NANO_SWITCH_B,
         "actions": [
-            TAP_TEMPO(
+            MORPH_BUTTON(
                 display = DISPLAY_FOOTER_2,
-                color = Colors.LIGHT_GREEN,
-                use_leds = False
-            ),
-            SHOW_TEMPO()    # Shows beats with the LED(s)
+                #color = "kemper"
+            )
         ]
     }
 ]
-
-######################################################################
-
-# Expression pedals and other inputs
-Inputs = None