+++ examples/Kemper Profiler Player/MIDICaptain Nano 4/tunetown KBDN/inputs.py
@@ -6,46 +6,55 @@
  
 from pyswitch.hardware.Hardware import Hardware
 
-#from pyswitch.misc import Colors
+from pyswitch.misc import Colors
+
 from pyswitch.clients.kemper import KemperEffectSlot
+
 from display import DISPLAY_HEADER_1, DISPLAY_HEADER_2, DISPLAY_FOOTER_1, DISPLAY_FOOTER_2
 
+from pyswitch.clients.kemper.actions.morph import MORPH_DISPLAY
+from pyswitch.clients.kemper.actions.tuner import TUNER_MODE
 from pyswitch.clients.kemper.actions.effect_state import EFFECT_STATE
+from pyswitch.clients.kemper.actions.rig_select import RIG_SELECT, RIG_SELECT_DISPLAY_TARGET_RIG
 
 
 # Defines the switch assignments
-Switches = [
+Inputs = [
 
     # Switch 1
     {
         "assignment": Hardware.PA_MIDICAPTAIN_NANO_SWITCH_1,
         "actions": [
-            EFFECT_STATE(
-                slot_id = KemperEffectSlot.EFFECT_SLOT_ID_A,
+            TUNER_MODE(
                 display = DISPLAY_HEADER_1
-            )                         
+            ),
+            MORPH_DISPLAY()        
         ]
     },
 
     # Switch 2
     {
         "assignment": Hardware.PA_MIDICAPTAIN_NANO_SWITCH_2,
-        "actions": [
+        "actions": [   
             EFFECT_STATE(
-                slot_id = KemperEffectSlot.EFFECT_SLOT_ID_B,
+                slot_id = KemperEffectSlot.EFFECT_SLOT_ID_X,
                 display = DISPLAY_HEADER_2
             )
-        ]
+        ]               
     },
 
     # Switch A
     {
         "assignment": Hardware.PA_MIDICAPTAIN_NANO_SWITCH_A,
         "actions": [
-            EFFECT_STATE(
-                slot_id = KemperEffectSlot.EFFECT_SLOT_ID_DLY,
-                display = DISPLAY_FOOTER_1
-            )
+            RIG_SELECT(
+                rig = 3,
+                rig_off = 1,
+                display = DISPLAY_FOOTER_1,
+                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG,
+                color = Colors.PINK,
+                text = "Synth-3"
+            )            
         ]
     },
     
@@ -53,15 +62,13 @@ Switches = [
     {
         "assignment": Hardware.PA_MIDICAPTAIN_NANO_SWITCH_B,
         "actions": [
-            EFFECT_STATE(
-                slot_id = KemperEffectSlot.EFFECT_SLOT_ID_REV,
-                display = DISPLAY_FOOTER_2
+            RIG_SELECT(
+                rig = 2,
+                rig_off = 1,
+                display = DISPLAY_FOOTER_2,
+                display_mode = RIG_SELECT_DISPLAY_TARGET_RIG,
+                text = "Lead-2"
             )
         ]
     }
 ]
-
-######################################################################
-
-# Expression pedals and other inputs
-Inputs = None